1.0.8

+++ job.sh
@@ -2,9 +2,14 @@
 
 . /chia-blockchain/venv/bin/activate
 chia stop all
+
 cd /root/.chia/mainnet/db/
-cp blockchain_v2_mainnet.sqlite /var/www/html/blockchain_v2_mainnet.sqlite.2
+test -f blockchain_v2_mainnet.sqlite && cp blockchain_v2_mainnet.sqlite /var/www/html/blockchain_v2_mainnet.sqlite.2
+test -f blockchain_v2_testnet10.sqlite && cp blockchain_v2_testnet10.sqlite /var/www/html/blockchain_v2_testnet10.sqlite.2
+
 cd /var/www/html
-mv blockchain_v2_mainnet.sqlite.2 blockchain_v2_mainnet.sqlite
+test -f blockchain_v2_mainnet.sqlite.2 && mv blockchain_v2_mainnet.sqlite.2 blockchain_v2_mainnet.sqlite
+test -f blockchain_v2_testnet10.sqlite.2 && mv blockchain_v2_testnet10.sqlite.2 blockchain_v2_testnet10.sqlite
+
 chown -R www-data:www-data /var/www/html
 chia start node
